v1.1 Update
Dependencies are pinned to ensure that they continue to work. There have
been breaking updates in some upstream projects, while others are
beginning to bit rot. For now I am preferring the path of least
resistance that will keep the module useful.

The module now generates a single image only. This is not configurable.

Renamed the -a/--artifacts CLI option to -d/--dalle. This is a breaking
change.

Dropped support for -v as a short-form of the CLI --version. This is a
minor breaking change.

Dropped all support for GPU-acceleration.

Added the -v/--vqgan and -o/--output CLI options. The pre-trained VQGAN
model should now be downloaded separately and provided on the --vqgan CLI
option. This is a breaking change.

The README demonstrates where pre-trained models can be downloaded from,
how to run the module, and how to build a Docker image and container to
run the module.